% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytometryQC-package.R
\docType{package}
\name{CytometryQC-package}
\alias{CytometryQC}
\alias{CytometryQC-package}
\title{CytometryQC: Tools for the Cytometry QC tracking}
\description{
To learn more about how to use CytometryQC, please start by browsing the vignettes:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidrach.github.io/CytometryQC/}
  \item Report bugs at \url{https://github.com/DavidRach/CytometryQC/issues}
}

}
\author{
\strong{Maintainer}: David Rach \email{drach@som.umaryland.edu} (\href{https://orcid.org/0000-0003-2235-5585}{ORCID})

}
\keyword{internal}
