% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoricalPages.R
\name{HistoricalPages}
\alias{HistoricalPages}
\title{Checks data folder, determines instruments, creates GitHub project for each,
 checks for historical data and creates a year page for each, wraps up by updating
 the main webpage links}
\usage{
HistoricalPages(githubusername = "umgccfcss")
}
\value{
Assembled Historical Websites
}
\description{
Checks data folder, determines instruments, creates GitHub project for each,
 checks for historical data and creates a year page for each, wraps up by updating
 the main webpage links
}
\examples{

A <- 2 + 2
}
