% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexUpdate.R
\name{IndexUpdate}
\alias{IndexUpdate}
\title{Triggered by AddInstrument, adds index.qmd chunks for that instrument.}
\usage{
IndexUpdate(outpath, name)
}
\description{
Triggered by AddInstrument, adds index.qmd chunks for that instrument.
}
