% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.R
\name{OperatingSystemCheck}
\alias{OperatingSystemCheck}
\title{Checks Operating System, sets appropiate file path}
\usage{
OperatingSystemCheck()
}
\value{
The Path to the Documents Folder
}
\description{
Checks Operating System, sets appropiate file path
}
