% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_404}
\alias{QMD_404}
\title{Creates the 404.qmd file}
\usage{
QMD_404(
  outpath,
  organization = "UMGCC FCSS Instrument",
  github_page = "umgccfcss.github.io"
)
}
\arguments{
\item{outpath}{The file.location to save the .qmd file to}
}
\value{
A 404.qmd file
}
\description{
Creates the 404.qmd file
}
