% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_Data}
\alias{QMD_Data}
\title{Creates generic Data.qmd file}
\usage{
QMD_Data(outpath)
}
\arguments{
\item{outpath}{The file.location to save the .qmd file to}
}
\value{
A generic Data.qmd file
}
\description{
Creates generic Data.qmd file
}
