% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_help}
\alias{QMD_help}
\title{Creates the help.qmd file}
\usage{
QMD_help(outpath)
}
\arguments{
\item{outpath}{The file.location to save the .qmd file to}
}
\value{
A help.qmd file
}
\description{
Creates the help.qmd file
}
