% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_Historical}
\alias{QMD_Historical}
\title{Creates generic Historical.qmd file}
\usage{
QMD_Historical(outpath)
}
\arguments{
\item{outpath}{The file.location to save the .qmd file to}
}
\value{
A generic Historical.qmd file
}
\description{
Creates generic Historical.qmd file
}
