% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_HistoricalInstrument}
\alias{QMD_HistoricalInstrument}
\title{Creates generic HistoricalInstument.qmd file}
\usage{
QMD_HistoricalInstrument(outpath, name, organization, organization_website)
}
\arguments{
\item{outpath}{The file.location to save the .qmd file to}
}
\value{
A generic HistoricalInstument.qmd file
}
\description{
Creates generic HistoricalInstument.qmd file
}
