% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_index}
\alias{QMD_index}
\title{Creates generic index.qmd file}
\usage{
QMD_index(outpath)
}
\arguments{
\item{outpath}{The file.location to save the .qmd file to}
}
\value{
A generic index.qmd file
}
\description{
Creates generic index.qmd file
}
