% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_Instrument}
\alias{QMD_Instrument}
\title{Creates generic Instrument.qmd file}
\usage{
QMD_Instrument(outpath)
}
\arguments{
\item{outpath}{The file.location to save the .qmd file to}
}
\value{
A generic instrument.qmd file
}
\description{
Creates generic Instrument.qmd file
}
