% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_README}
\alias{QMD_README}
\title{Creates generic README.md file}
\usage{
QMD_README(outpath, organization, organization_website)
}
\arguments{
\item{outpath}{The file.location to save the .qmd file to}
}
\value{
A generic README.md file
}
\description{
Creates generic README.md file
}
