% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructureQMD.R
\name{QMD_yaml}
\alias{QMD_yaml}
\title{Creates initial quarto yaml}
\usage{
QMD_yaml(
  outpath,
  organization = "UMGCC FCSS",
  github_page = "umgccfcss.github.io",
  institution = "University of Maryland, Baltimore"
)
}
\arguments{
\item{outpath}{The file.location to save the .yaml file to}
}
\value{
The .yaml file needed to generate the website
}
\description{
Creates initial quarto yaml
}
