% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standalone_Chorus.R
\name{Standalone_Chorus}
\alias{Standalone_Chorus}
\title{Build a standalone offline webpage to visualize QC from BD instruments
that run the Chorus software. Please note, it is not yet tied in to the 
rest of the CytometryQC infrastructure, as this was a spontaneous
Cytometry Discord fun project.}
\usage{
Standalone_Chorus(
  input,
  organization_name = "UMGCC FCSS",
  institution_name = "University of Maryland, Baltimore",
  FolderName = "QC_Chorus"
)
}
\arguments{
\item{organization_name}{Default is "UMGCC FCSS", provide your own organization
name within " " for it to appear on the offline webpage}

\item{institution_name}{Default is "University of Maryland, Baltimore", provide
your own institutions name within " " for it to appear on the offline webpage}

\item{FolderName}{Default "QC_Chorus", the folder name under Documents where to 
keep this project.}

\item{inpath}{A file path to the folder that contains the Chorus QC.pdf files.}
}
\description{
Build a standalone offline webpage to visualize QC from BD instruments
that run the Chorus software. Please note, it is not yet tied in to the 
rest of the CytometryQC infrastructure, as this was a spontaneous
Cytometry Discord fun project.
}
