% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddInstruments.R
\name{AddInstruments}
\alias{AddInstruments}
\title{Adds Instruments to the base Webpage}
\usage{
AddInstruments(
  name,
  manufacturer = "Cytek",
  uv = 16,
  violet = 16,
  blue = 14,
  yellowgreen = 10,
  red = 8,
  organization_name = NULL,
  organization_website = NULL,
  githubusername = NULL,
  TheFCSFolderPath = NULL,
  CytekbioExportFolderPath = NULL
)
}
\arguments{
\item{name}{Desired name for the instrument}

\item{manufacturer}{Instrument Manufacturer}

\item{uv}{Number of UV detectors}

\item{violet}{Number of violet detectors}

\item{blue}{Number of blue detectors}

\item{yellowgreen}{Number of yellow-green detectors}

\item{red}{Number of red detectors}

\item{organization_name}{Name of the organization, default NULL defaults to 
value provided during FolderSetup}

\item{organization_website}{Link to organization website, default NULL defaults
to value provided during FolderSetup}

\item{githubusername}{GitHub user name, default NULL defaults to value provided 
during FolderSetup}

\item{TheFCSFolderPath}{Default NULL sets path to Cytek file.path 
'C:/CytekbioExport/Setup/DailyQC', for alternate file locations or manufacturers, please
provide a file.path to the fcs folder location}

\item{CytekbioExportFolderPath}{Default NULL sets path to Cytek file.path 'C:/CytekbioExport',
for alternate locations of the CytekbioExport folder, please provide a file path.}
}
\value{
Updated webpage
}
\description{
Adds Instruments to the base Webpage
}
\examples{

AddInstruments(name="5L", manufacturer="Cytek", uv=16, violet=16, blue=14,
yellowgreen=10, red=8, TheFCSFolderPath="C:/CytekbioExport/Setup/DailyQC")

}
