% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddInstruments.R
\name{AddInstruments}
\alias{AddInstruments}
\title{Adds Instruments to the base Webpage}
\usage{
AddInstruments(
  name,
  manufacturer,
  uv = 16,
  violet = 16,
  blue = 14,
  yellowgreen = 10,
  red = 8
)
}
\arguments{
\item{outpath}{}
}
\value{
Updated webpage
}
\description{
Adds Instruments to the base Webpage
}
