% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddInstruments.R
\name{AddInstruments}
\alias{AddInstruments}
\title{Adds Instruments to the base Webpage}
\usage{
AddInstruments(
  name,
  manufacturer = "Cytek",
  uv = 16,
  violet = 16,
  blue = 14,
  yellowgreen = 10,
  red = 8,
  organization = "UMGCC FCSS",
  githubusername = "UMGCCFCSS",
  organization_website = "https://www.medschool.umaryland.edu/cibr/core/umgccc_flow/"
)
}
\arguments{
\item{name}{Desired name for the instrument}

\item{manufacturer}{Instrument Manufacturer}

\item{uv}{Number of UV detectors}

\item{violet}{Number of violet detectors}

\item{blue}{Number of blue detectors}

\item{yellowgreen}{Number of yellow-green detectors}

\item{red}{Number of red detectors}

\item{organization}{Name of the organization to include on the page}

\item{githubusername}{GitHub user name, used for path setting.}

\item{organization_website}{Link to organization website to include on the page}
}
\value{
Updated webpage
}
\description{
Adds Instruments to the base Webpage
}
\examples{

A <- 2 + 2

}
