% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderSetup.R
\name{FolderSetup}
\alias{FolderSetup}
\title{Checks for an existing InstrumentQC folder, if not present creates the folder
and populates it with the required website components}
\usage{
FolderSetup(
  organization_name = "UMGCCC FCSR",
  organization_website = "https://www.medschool.umaryland.edu/cibr/core/umgccc_flow/",
  githubusername = "umgcccfcsr",
  institution_name = "University of Maryland, Baltimore",
  AlternateDirectory = NULL,
  FolderName = "InstrumentQC2",
  SetUpGit = FALSE
)
}
\arguments{
\item{organization_name}{The organization name, ex. UMGCC FCSS}

\item{organization_website}{The organizations website, ex. https://www.medschool.umaryland.edu/cibr/core/umgccc_flow/}

\item{githubusername}{The GitHub user name, ex. umgccfcss}

\item{institution_name}{The institution name, ex. University of Maryland, Baltimore}

\item{AlternateDirectory}{Provide a file path if desire to save somewhere not Documents folder.}

\item{FolderName}{Default InstrumentQC2, sets folder/repository name}

\item{SetUpGit}{Default is FALSE, when git token credentials are present, 
it will generate a git repository for the folder and push to GitHub.}
}
\value{
A generalized InstrumentQC folder to which additional elements can be added
}
\description{
Checks for an existing InstrumentQC folder, if not present creates the folder
and populates it with the required website components
}
\examples{

tmp <- withr::local_tempdir(pattern = "CytometryQC")
 FolderSetup(
 organization_name = "UMGCCC FCSR",
 organization_website = "https://www.medschool.umaryland.edu/cibr/core/umgccc_flow/",
 githubusername = "umgcccfcsr",
 institution_name = "University of Maryland, Baltimore",
 AlternateDirectory = tmp # For Testing Only
 )


}
