% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderSetup.R
\name{FolderSetup}
\alias{FolderSetup}
\title{Checks for an existing InstrumentQC folder, if not present creates the folder
and populates it with the required website components}
\usage{
FolderSetup(
  SetUpGit = FALSE,
  organization_name = "UMGCC FCSS",
  organization_website = "https://www.medschool.umaryland.edu/cibr/core/umgccc_flow/",
  githubusername = "umgccfcss",
  institution_name = "University of Maryland, Baltimore",
  FolderName = "InstrumentQC2",
  AlternateDirectory = NULL
)
}
\arguments{
\item{SetUpGit}{Default is FALSE, when git token credentials are present, 
it will generate a git repository for the folder and push to GitHub.}

\item{organization_name}{The organization name, ex. UMGCC FCSS}

\item{organization_website}{The organizations website, ex. https://www.medschool.umaryland.edu/cibr/core/umgccc_flow/}

\item{githubusername}{The GitHub user name, ex. umgccfcss}

\item{institution_name}{The institution name, ex. University of Maryland, Baltimore}

\item{FolderName}{Default InstrumentQC2, sets folder/repository name}

\item{AlternateDirectory}{Provide a file path if desire to save somewhere not Documents folder.}
}
\value{
A generalized InstrumentQC folder to which additional elements can be added
}
\description{
Checks for an existing InstrumentQC folder, if not present creates the folder
and populates it with the required website components
}
\examples{

\dontrun{
FolderSetup()
}

}
