% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.R
\name{FolderSetup}
\alias{FolderSetup}
\title{Checks for existing InstrumentQC folder, if none present
creates a new one and populates with the correct folders}
\usage{
FolderSetup(SetUpGit = FALSE)
}
\value{
If folder not present, creates folder under user Documents
folder
}
\description{
Checks for existing InstrumentQC folder, if none present
creates a new one and populates with the correct folders
}
\examples{

A <- 2 + 2

}
