% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderWebsite.R
\name{RenderWebsite}
\alias{RenderWebsite}
\title{Quarto renders the project to produce the website, checks for git, sends to GitHub.}
\usage{
RenderWebsite()
}
\value{
Rendered .html files containing the website to the docs folder, pushes changes
to GitHub
}
\description{
Quarto renders the project to produce the website, checks for git, sends to GitHub.
}
\examples{

A <- 2 + 2

}
